+++ apps/blog-app/config/jest/app-test-providers.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { AppProviders } from '../../src/app-providers';
+
+export const AppTestProviders: React.FC = ({ children }) => {
+  return <AppProviders>{children}</AppProviders>;
+};