+++ apps/blog-app/config/jest/test-utils.tsx
@@ -4,15 +4,11 @@
  */
 import React from 'react';
 import { render } from '@testing-library/react';
-import { AppProviders } from '../../src/app-providers';
-
-const AllTheProviders: React.FC = ({ children }) => {
-  return <AppProviders>{children}</AppProviders>;
-};
+import { AppTestProviders } from './app-test-providers';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const customRender = (ui: React.ReactElement, options?: any) =>
-  render(ui, { wrapper: AllTheProviders, ...options });
+  render(ui, { wrapper: AppTestProviders, ...options });
 
 // re-export everything
 export * from '@testing-library/react';